--- yuaibin.txt
@@ -1,4 +1,0 @@
-Try to understand the instructions.
-but find it a little bit difficult.
-orz
-from Edwin Yu(yuaibin)
